+++ frontend/src/components/TopThreeStudiosWithWinners.jsx
